cortex-engine/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cortex-engine"
version = "0.1.0"
description = "CortexOS attribution engine – EAS fast-path backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.30,<1",
    "pydantic>=2.7,<3",
    "pydantic-settings>=2.3,<3",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.29,<1",
    "numpy>=1.26,<3",
    "sentence-transformers>=3.0,<4",
    "httpx>=0.27,<1",
    "opentelemetry-api>=1.25,<2",
    "opentelemetry-sdk>=1.25,<2",
    "alembic>=1.13,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "httpx",
]

[tool.setuptools.packages.find]
include = ["cortex*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
